update user editing

+++ src/main/java/com/humanresource/hr/role/RoleService.java
@@ -1,5 +1,7 @@
 package com.humanresource.hr.role;
 
+import com.humanresource.hr.helper.Constants;
+import com.humanresource.hr.helper.DeleteResponse;
 import com.humanresource.hr.permission.Permission;
 import com.humanresource.hr.permission.PermissionRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,26 +18,26 @@ public class RoleService {
     @Autowired
     private PermissionRepository permissionRepository;
 
-    public List<Role> fetchAllRoles(){
+    public List<Role> fetchAllRoles() {
         try {
             return roleRepository.findAll();
-        } catch (Exception e){
+        } catch (Exception e) {
             throw new RuntimeException(e.getMessage());
         }
     }
 
-    public Role createRole(Role role){
+    public Role createRole(Role role) {
         try {
             return roleRepository.save(role);
-        } catch (Exception e){
+        } catch (Exception e) {
             throw new RuntimeException(e.getMessage());
         }
     }
 
-    public Optional<Role> findOneRole(Long roleId){
-        try{
+    public Optional<Role> findOneRole(Long roleId) {
+        try {
             return roleRepository.findById(roleId);
-        } catch (Exception e){
+        } catch (Exception e) {
             throw new RuntimeException(e.getMessage());
         }
     }
@@ -43,42 +45,38 @@ public class RoleService {
     public Role assignPermissionsToRole(Long roleID, Long permID) {
 
         try {
-
             Set<Permission> permissions = null;
             Optional<Role> role = roleRepository.findById(roleID);
             Optional<Permission> permission = permissionRepository.findById(permID);
 
-        if(role.isPresent() && permission.isPresent()){
-            permissions = role.get().getPermissions();
-            permissions.add(permission.get());
-            role.get().setPermissions(permissions);
-            return roleRepository.save(role.get());
-        } else {
-            return null;
-        }
-        } catch (Exception e){
+            if (role.isPresent() && permission.isPresent()) {
+                permissions = role.get().getPermissions();
+                permissions.add(permission.get());
+                role.get().setPermissions(permissions);
+                return roleRepository.save(role.get());
+            } else {
+                return null;
+            }
+        } catch (Exception e) {
             throw new RuntimeException(e.getMessage());
         }
     }
 
-    public Map<String, String> deleteRole(Long roleID) {
-        Map<String,String> response = new HashMap<String, String>();
-
-        try{
+    public DeleteResponse deleteRole(Long roleID) {
+        DeleteResponse response = new DeleteResponse();
+        if (roleRepository.existsById(roleID)) {
             roleRepository.deleteById(roleID);
-            response.put("response", "Deleted successfully");
-            response.put("success", "true");
-            return response;
-
-        } catch (Exception e){
-            response.put("response", e.getMessage());
-            return response;
+            response.setMessage(Constants.DELETED_SUCCESSFULLY);
+            response.setSuccess(true);
+        } else {
+            response.setMessage(Constants.ENTITY_NOT_FOUND);
+            response.setSuccess(false);
         }
+        return response;
     }
 
     public Role fetchRole(Long roleID) {
-        Optional<Role> response = roleRepository.findById(roleID);
-        return response.orElse(null);
+        return roleRepository.findById(roleID).orElse(null);
     }
 
     public Role updateRole(Long roleID, Role role) {
@@ -88,7 +86,7 @@ public class RoleService {
             assert currentData != null;
             currentData.setName(role.getName());
             return roleRepository.save(currentData);
-        } catch (Exception e){
+        } catch (Exception e) {
             throw new RuntimeException(e.getMessage());
         }
     }
